Week 3

+++ src/main/java/dat3/car/api/CarController.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 import org.springframework.web.bind.annotation.*;
 
-import dat3.car.dto.CarRequest;
-import dat3.car.dto.CarResponse;
+import dat3.car.dto.car.CarRequest;
+import dat3.car.dto.car.CarResponse;
 import dat3.car.service.CarService;
 
 @RestController
@@ -26,7 +26,7 @@ public class CarController {
 
     // Role: ANONYMOUS
     @GetMapping("/{id}")
-    public CarResponse find(@PathVariable("id") long id) {
+    public CarResponse find(@PathVariable("id") int id) {
         return carService.find(id);
     }
 
@@ -44,7 +44,7 @@ public class CarController {
 
     // Role: ADMIN
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable("id") long id) {
+    public void delete(@PathVariable("id") int id) {
         carService.delete(id);
     }
 }
